+++ src/main/java/com/mgmtp/radio/sdo/SubscriptionEvents.java
@@ -4,6 +4,5 @@ public enum  SubscriptionEvents {
     register,
     forgot_password,
     reset_password,
-    song_history,
-    skip_rule
+    song_history
 }
